move logging pass logic to base class, add GolangLogPass
