+++ cloudpg/src/main/java/io/clouditor/graph/passes/python/PythonLogPass.kt
@@ -1,0 +1,28 @@
+package io.clouditor.graph.passes.python
+
+import de.fraunhofer.aisec.cpg.TranslationResult
+import de.fraunhofer.aisec.cpg.graph.Node
+import de.fraunhofer.aisec.cpg.graph.statements.expressions.MemberCallExpression
+import de.fraunhofer.aisec.cpg.processing.IVisitor
+import de.fraunhofer.aisec.cpg.processing.strategy.Strategy
+import io.clouditor.graph.*
+import io.clouditor.graph.passes.LogPass
+
+class PythonLogPass : LogPass() {
+    override fun accept(t: TranslationResult) {
+        // if (this.lang is PythonLanguageFrontend) {
+        for (tu in t.translationUnits) {
+            tu.accept(
+                Strategy::AST_FORWARD,
+                object : IVisitor<Node?>() {
+                    fun visit(m: MemberCallExpression) {
+                        if (m.name == "info" && m.base.name == "logging") {
+                            handleLog(t, m, tu)
+                        }
+                    }
+                }
+            )
+        }
+    }
+
+}
